Move To Active Job
Solidus and Spree3 all require fairly new versions of rails. Assume that
Active Job is available and will be used to queue messages due to best
practices. Solidus branch will only support the use of ActiveJob. Even
if folks don't want to use an external DB to manage the queueing, a
  delayed job adapter exists and will work wonderfully with Psql/Mysql.