Add new shipment mailer
Instead of sending an email per order, we have changed it so we send an
email PER shipment.

From the client:
6) Since there can be multiple shipments (and tracking numbers), please
ensure that the order shipped confirmation email only include items from
that particular shipment and not for the entire order, and that for
multiple shipments the customer would receive multiple shipment
confirmation emails.